Add DropEncryptRuleBackendHandler
